Removes offenders from summary if no sentence start date
When an offender is released, their sentence_start_date is set to nil,
so this will cause a crash in the summary page because we can't
determine their responsibility.

This PR will exclude the offender from get_offenders_for_prison by
claiming they are not sentenced if they have no sentence_start_date.