Handle missing primary_pom_allocated_at
Sometimes, in an active allocation we have a missing
primary_pom_allocated_at for historical reasons.  This causes to_date to
fail when trying to find out when an allocation happened (so it can be
assigned to the offender).

In these cases, we will use the updated_at date as the allocation date,
although more thoroughly we should probably iterate through the history
to find the most recent primary pom allocated at. We are not doing this
as we hope to fix up historical data by finding current allocation
events with a missing primary_pom_allocated_at and update them.