Last_known_address partial now works due to the addition of aria_controls and aria_expanded being added to the form.
Added a virtual attribute :last_known_address to the case_information
model for validation to the yes/no radio buttons for whether the
offender's last known address was in Scotland, Wales or Northern
Ireland. Clicking yes will allow for the the user to populate
:probation_service with either 'Scotland, Wales or Northern Ireland'.

If the offender's last known address is Scotland or Northern Ireland
then we will need to populate the tier with the value 'N/A'.The
validation for the tier has been updated to reflect this.