Update responsibility & recommendation services
We have implemented a check as to whether an offender is Scottish or Northern Irish
and this means that we have been able to update the responsibility and recommendation service to
make use of these.  If the offender is Scottish or Northern Irish then they will be recommended a
Prison POM at allocation, and once allocated the POM will show as supporting/community.

We have also started to update the feature specs to reflect the changes to the page design/layout, but have not yet
done this for the England/Wales path.