Refactor the hide and show functionality into functions and use them in the edit form by listening for the on_click event and then firing the hide or show function.
We now can call the turbolinks:load event which will only be actioned
within the case_information edit page following changes made to
application.html.erb. If we are in the case_information edit page, once
the page has loaded we will hide the rest of the form elements if the
Scotland or Northern Ireland radio buttons have been checked.