Add failing acceptance spec for missing data
In the wake of 3a3ec5 (which updates how the recommended
responsibility is calculated), we discovered issues in production with
certain offenders. These offenders, when looked up in the NOMIS API,
were missing certain dates the app always expects to be present.

When this data isn't present, the app 'crashes' (returning a 500 for
the request, but continuing to serve other requests). This is the case
for pages with multiple offenders if only a single one is missing data.

This commit does not stop the app crashing - it only adds a spec where
the absolute minimum data is returned from the API, to start working
through the errors.